requirements.txt
python-socketio
Flask~=3.1.2
Flask-SocketIO~=5.5.1
twitchAPI~=4.5.0
python-dotenv~=1.1.1
eventlet~=0.40.3
requests~=2.32.5